requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
celery==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.6.0
crispy-tailwind==1.0.3
Django==5.0.7
django-allauth==0.63.6
django-allauth-ui==1.2.0
django-crispy-forms==2.3
django-dotenv==1.4.2
django-filter==24.2
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
gunicorn==22.0.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
rapper==0.1.1
rdflib==7.0.0
rdfpandas==1.1.6
redis==5.0.7
requests==2.32.3
rich==13.7.1
six==1.16.0
slippers==0.6.2
soupsieve==2.5
sqlparse==0.5.1
text-unidecode==1.3
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
